# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VTEI Student API",
        "description": "REST API для мобільного додатку студентів ВТЕІ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/voting": {
            "get": {
                "tags": [
                    "Voting"
                ],
                "summary": "Кандидати голосування",
                "description": "Повертає список кандидатів студентського самоврядування та чи вже проголосував студент.",
                "operationId": "58bddeb5b0e341b3c7e4d47e37b46e3e",
                "responses": {
                    "200": {
                        "description": "Кандидати і статус голосування"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/voting/vote": {
            "post": {
                "tags": [
                    "Voting"
                ],
                "summary": "Проголосувати",
                "description": "Записує голос студента за кандидата. Повторне голосування неможливе.",
                "operationId": "6d260cab226642395511abaae4f38621",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "candidate_student_id"
                                ],
                                "properties": {
                                    "candidate_student_id": {
                                        "description": "student_grupa_id кандидата",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Голос зараховано"
                    },
                    "409": {
                        "description": "Вже проголосував"
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/voting/results": {
            "get": {
                "tags": [
                    "Voting"
                ],
                "summary": "Результати голосування",
                "description": "Повертає кількість голосів за кожного кандидата.",
                "operationId": "655f0294e2683d0da0adf8299298a3a7",
                "responses": {
                    "200": {
                        "description": "Результати"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}